fix(auth): maintain session after page refresh in AuthContext
